% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-igraph.r
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.Rcpp_SimplexTree}
\alias{as_igraph.simplextree}
\alias{as_igraph.igraph}
\alias{as_igraph.network}
\title{Convert objects to class 'igraph'}
\usage{
as_igraph(x, ...)

\method{as_igraph}{default}(x, ...)

\method{as_igraph}{Rcpp_SimplexTree}(x, ...)

\method{as_igraph}{simplextree}(x, ...)

\method{as_igraph}{igraph}(x, ...)

\method{as_igraph}{network}(x, ...)
}
\arguments{
\item{x}{An R object to be coerced. See Details.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class 'igraph'.
}
\description{
This generic function...
}
\examples{
# convert a TDA filtration object
t <- 2 * pi * c(0, 1, 3, 6) / 7
rf <- TDA::ripsFiltration(
  cbind(x = cos(t), y = sin(t)),
  maxdimension = 2L, maxscale = 1.7
)
print(rf$cmplx)
ig_rf <- as_igraph(rf)
print(ig_rf)
ig_rf2 <- as_igraph(rf$cmplx)
print(ig_rf2)

# convert a simplextree object
st <- simplextree::simplex_tree()
st$insert(list(1:3, 4:5, 6))
ig_st <- as_igraph(st)
print(ig_st)

# convert an igraph object
g <- igraph::graph(c(1,2, 2,3, 1,3, 3,4))
print(g)
ig_g <- as_igraph(g)
print(ig_g)

# convert a network object
el <- data.frame(tails = c(1, 2, 1, 3), heads = c(2, 3, 3, 4))
n <- network::network.edgelist(el, network::network.initialize(4))
print(n)
ig_n <- as_igraph(n)
print(ig_n)
}
