% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-network.r
\name{as_network}
\alias{as_network}
\alias{as_network.default}
\alias{as_network.Rcpp_SimplexTree}
\alias{as_network.simplextree}
\alias{as_network.igraph}
\alias{as_network.network}
\title{Convert objects to class 'network'}
\usage{
as_network(x, ...)

\method{as_network}{default}(x, index = NULL, ...)

\method{as_network}{Rcpp_SimplexTree}(x, index = NULL, ...)

\method{as_network}{simplextree}(x, index = NULL, ...)

\method{as_network}{igraph}(x, ...)

\method{as_network}{network}(x, ...)
}
\arguments{
\item{x}{An R object to be coerced. See Details.}

\item{...}{Additional arguments passed to methods.}

\item{index}{Character string to be added as a vertex attribute containing
0-simplex indices. Ignored if \code{NULL} (the default).}
}
\value{
An object of class 'network'.
}
\description{
This generic function...
}
\examples{
# convert a TDA filtration object
t <- 2 * pi * c(0, 1, 3, 6) / 7
rf <- TDA::ripsFiltration(
  cbind(x = cos(t), y = sin(t)),
  maxdimension = 2L, maxscale = 1.7
)
print(rf$cmplx)
nw_rf <- as_network(rf)
print(nw_rf)
nw_rf2 <- as_network(rf$cmplx)
print(nw_rf2)

# convert a simplextree object
st <- simplextree::simplex_tree()
st$insert(list(3:5, 5:6, 8))
nw_st <- as_network(st)
print(nw_st)

# convert an igraph object
g <- igraph::graph(c(1,2, 2,3, 1,3, 3,4))
print(g)
nw_g <- as_network(g)
print(nw_g)
}
