% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplex-package.r
\docType{package}
\name{interplex}
\alias{interplex}
\title{\strong{interplex} package}
\description{
This is a helper package to convert simplicial complexes between
different data structures.
}
\details{
This package helps interface between different topological data analytic
packages and workflows by converting simplicial complex objects stored using
different data structures. Each conversion is designed to retain as much
annotation as possible, whether of simplices or of the complex.

The package supports conversions between simplicial complexes stored using
the following data structures:
\itemize{
\item a complete list of simplices,
as stored in \code{cmplx} values of filtration objects
in \link[TDA:TDA-package]{the TDA package}
\item an object of class 'Rcpp_SimplexTree' or 'simplextree'
as implemented
in \link[simplextree:simplextree-package]{the simplextree package}
\item an 'igraph' object,
as implemented
in \link[=igraph]{the igraph package}
\item a 'network' object,
as implemented
in \link[network:network-package]{the network package}
}

When converting from a simplicial complex structure to a graph structure,
only the 1-skeleton is converted; simplices of dimension > 1 are discarded
and no warning is printed.
}
