% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-simplextree.r
\name{as_simplextree}
\alias{as_simplextree}
\alias{as_simplextree.default}
\alias{as_simplextree.igraph}
\alias{as_simplextree.network}
\title{Convert objects to class "simplextree"}
\usage{
as_simplextree(x, ...)

\method{as_simplextree}{default}(x, ...)

\method{as_simplextree}{igraph}(x, ...)

\method{as_simplextree}{network}(x, ...)
}
\arguments{
\item{x}{An R object to be coerced. See Details.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class "simplextree".
}
\description{
This generic function...
}
\examples{
# convert a TDA filtration object
rf <- TDA::ripsFiltration(TDA::circleUnif(4L), maxdimension = 2L, maxscale = 1)
print(rf$cmplx)
st_rf <- as_simplextree(rf)
plot(st_rf)
st_rf2 <- as_simplextree(rf$cmplx)
plot(st_rf2)

# convert an igraph object
g <- igraph::graph(c(1,2, 2,3, 1,3, 3,4))
plot(g)
st_g <- as_simplextree(g)
plot(st_g)

# convert a network object
el <- data.frame(tails = c(1, 2, 1, 3), heads = c(2, 3, 3, 4))
n <- network::network.edgelist(el, network::network.initialize(4))
plot(n)
st_n <- as_simplextree(n)
plot(st_n)
}
