% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplex-package.r
\docType{package}
\name{interplex}
\alias{interplex}
\title{\strong{interplex} package}
\description{
This is a helper package to convert simplicial complexes between
different data structures.
}
\details{
This package helps interface between different topological data
analytic packages and workflows by converting simplicial complex objects
stored using different data structures. Each conversion is designed to
retain as much annotation as possible, whether of simplices or of the
complex.
}
